wm_chess/pyproject.toml
[project]
name = "wm-chess"
version = "0.1.0"
description = "Shared chess core (board encoder, ResNet, MCTS, arena) for world-models experiments"
requires-python = ">=3.12,<3.13"
dependencies = [
    "torch>=2.5",
    "numpy>=1.26",
    "python-chess>=1.999",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wm_chess"]
